+++ riven_/update.py
@@ -158,14 +158,7 @@ class RivenUpdate(Update):
                 time.sleep(10)
             env_file_path = os.path.join(config_dir, "src", ".env")
             load_dotenv(env_file_path)
-            venv_path, poetry_executable = setup_poetry_environment(
-                self.process_handler, config_dir
-            )
-            if not venv_path or not poetry_executable:
-                self.logger.error(
-                    f"Failed to set up Poetry environment for {process_name}"
-                )
-                return
+            venv_path = os.path.join(config_dir, "venv")
             riven_python = os.path.join(venv_path, "bin", "python")
             command = [riven_python, "src/main.py"]
             env = os.environ.copy()