+++ utils/processes.py
@@ -10,29 +10,19 @@ class ProcessHandler:
         if cls._instance is None:
             cls._instance = super(ProcessHandler, cls).__new__(cls)
             cls._instance.init_attributes(*args, **kwargs)
-            signal.signal(signal.SIGCHLD, cls._instance.reap_zombies)
+            signal.signal(signal.SIGTERM, cls._instance.shutdown)
+            signal.signal(signal.SIGINT, cls._instance.shutdown)
         return cls._instance
 
     def init_attributes(self, logger):
         self.logger = logger
         self.processes = {}
+        self.process_names = {}
         self.subprocess_loggers = {}
         self.stdout = ""
         self.stderr = ""
         self.returncode = None
-
-    def __init__(self, logger):
-        pass
-
-    def reap_zombies(self, signum, frame):
-        while True:
-            try:
-                pid, _ = os.waitpid(-1, os.WNOHANG)
-                if pid == 0:
-                    break
-                self.logger.info(f"Reaped zombie process with PID: {pid}")
-            except ChildProcessError:
-                break
+        self.shutting_down = False
 
     def start_process(
         self,
@@ -44,32 +34,15 @@ class ProcessHandler:
         env=None,
     ):
         try:
-            try:
-                pwd.getpwuid(user_id)
-            except KeyError:
-                self.logger.error(
-                    f"UID {user_id} does not exist. Process will not start."
-                )
-                return None
-
-            try:
-                grp.getgrgid(group_id)
-            except KeyError:
-                self.logger.error(
-                    f"GID {group_id} does not exist. Process will not start."
-                )
-                return None
 
             def preexec_fn():
                 os.setgid(group_id)
                 os.setuid(user_id)
 
-            if key_type is not None:
-                self.logger.info(f"Starting {process_name} w/ {key_type} subprocess")
-                process_description = f"{process_name} w/ {key_type}"
-            else:
-                self.logger.info(f"Starting {process_name} subprocess")
-                process_description = f"{process_name}"
+            process_description = (
+                f"{process_name} w/ {key_type}" if key_type else process_name
+            )
+            self.logger.info(f"Starting {process_description} process")
 
             if isinstance(command, str):
                 command = shlex.split(command)
@@ -78,43 +51,36 @@ class ProcessHandler:
             if env:
                 process_env.update(env)
 
-            if process_name in [
-                "rclone",
-                "poetry_install",
-                "install_poetry",
-                "poetry_env_setup",
-                "PostgreSQL_init",
-                "npm_install",
-                "node_build",
-                "python_env_setup",
-                "install_requirements",
-                "setup_env_and_install",
-                "dotnet_env_restore",
-                "dotnet_publish_api",
-                "dotnet_publish_scraper",
-            ]:
-                process = subprocess.Popen(
-                    command,
-                    stdout=subprocess.PIPE,
-                    stderr=subprocess.PIPE,
-                    start_new_session=True,
-                    cwd=config_dir,
-                    universal_newlines=True,
-                    bufsize=1,
-                    env=process_env,
-                )
-            else:
-                process = subprocess.Popen(
-                    command,
-                    stdout=subprocess.PIPE,
-                    stderr=subprocess.PIPE,
-                    start_new_session=True,
-                    cwd=config_dir,
-                    universal_newlines=True,
-                    bufsize=1,
-                    preexec_fn=preexec_fn,
-                    env=process_env,
-                )
+            process = subprocess.Popen(
+                command,
+                stdout=subprocess.PIPE,
+                stderr=subprocess.PIPE,
+                start_new_session=True,
+                cwd=config_dir,
+                universal_newlines=True,
+                bufsize=1,
+                preexec_fn=(
+                    preexec_fn
+                    if process_name
+                    not in [
+                        "rclone",
+                        "poetry_install",
+                        "install_poetry",
+                        "poetry_env_setup",
+                        "PostgreSQL_init",
+                        "npm_install",
+                        "node_build",
+                        "python_env_setup",
+                        "install_requirements",
+                        "setup_env_and_install",
+                        "dotnet_env_restore",
+                        "dotnet_publish_api",
+                        "dotnet_publish_scraper",
+                    ]
+                    else None
+                ),
+                env=process_env,
+            )
 
             if not suppress_logging:
                 self.subprocess_logger = SubprocessLogger(
@@ -126,7 +92,12 @@ class ProcessHandler:
                 )
 
             self.logger.info(f"{process_name} process started with PID: {process.pid}")
-            self.processes[process_name] = process
+            self.processes[process.pid] = {
+                "name": process_name,
+                "description": process_description,
+                "process_obj": process,
+            }
+            self.process_names[process_name] = process
             return process
 
         except Exception as e:
@@ -135,14 +106,36 @@ class ProcessHandler:
             )
             return None
 
+    def reap_zombies(self, signum, frame):
+        while True:
+            try:
+                pid, _ = os.waitpid(-1, os.WNOHANG)
+                if pid == 0:
+                    break
+                process_info = self.processes.pop(pid, {"description": "Unknown"})
+                process_name = process_info.get("name")
+                if process_name in self.process_names:
+                    del self.process_names[process_name]
+                self.logger.debug(
+                    f"Reaped zombie process with PID: {pid}, "
+                    f"Description: {process_info.get('description', 'Unknown')}"
+                )
+            except ChildProcessError:
+                break
+
     def wait(self, process_name):
-        process = self.processes.get(process_name)
+        if self.shutting_down:
+            self.logger.debug(f"Skipping wait for {process_name} due to shutdown mode.")
+            return
+        process = self.process_names.get(process_name)
         if process:
             try:
-                self.stdout, self.stderr = process.communicate()
+                process.wait()
                 self.returncode = process.returncode
-                self.stdout = self.stdout.strip() if self.stdout else ""
-                self.stderr = self.stderr.strip() if self.stderr else ""
+                if process.stdout:
+                    self.stdout = process.stdout.read().strip()
+                if process.stderr:
+                    self.stderr = process.stderr.read().strip()
             except Exception as e:
                 self.logger.error(
                     f"Error while waiting for process {process_name}: {e}"
@@ -151,7 +144,8 @@ class ProcessHandler:
                 if self.subprocess_loggers.get(process_name):
                     self.subprocess_loggers[process_name].stop_logging_stdout()
                     self.subprocess_loggers[process_name].stop_monitoring_stderr()
-                del self.processes[process_name]
+                del self.process_names[process_name]
+                del self.processes[process.pid]
         else:
             self.logger.error(f"No process found with the name {process_name}.")
 
@@ -162,14 +156,32 @@ class ProcessHandler:
             )
             self.logger.info(f"Stopping {process_description}")
 
-            process = self.processes.get(process_name)
+            process = self.process_names.get(process_name)
+            self.logger.debug(f"Process {process_name} found: {process}")
             if process:
                 process.terminate()
+                time.sleep(2)
                 try:
-                    process.wait(timeout=10)
-                    self.logger.info(
-                        f"{process_description} process terminated gracefully."
-                    )
+                    for attempt in range(3):
+                        self.logger.debug(
+                            f"Waiting for {process_description} to terminate (attempt {attempt + 1})..."
+                        )
+                        process.wait(timeout=10)
+                        self.logger.debug(f"Wating on process: {process}")
+                        if process.poll() is None:
+                            self.logger.info(
+                                f"{process_description} process terminated gracefully."
+                            )
+                            break
+                    else:
+                        self.logger.warning(
+                            f"{process_description} process did not terminate gracefully, forcing stop."
+                        )
+                        process.kill()
+                        process.wait()
+                        self.logger.info(
+                            f"{process_description} process killed forcefully."
+                        )
                 except subprocess.TimeoutExpired:
                     self.logger.warning(
                         f"{process_description} process did not terminate gracefully, forcing stop."
@@ -185,7 +197,12 @@ class ProcessHandler:
                     self.subprocess_loggers[process_name].stop_monitoring_stderr()
                     del self.subprocess_loggers[process_name]
 
-                del self.processes[process_name]
+                self.process_names.pop(process_name, None)
+                process_info = self.processes.pop(process.pid, None)
+                if process_info:
+                    self.logger.debug(
+                        f"Removed process {process_name} with PID {process.pid} from tracking."
+                    )
             else:
                 self.logger.warning(
                     f"{process_description} process not found or already stopped."
@@ -194,3 +211,40 @@ class ProcessHandler:
             self.logger.error(
                 f"Error stopping subprocess for {process_description}: {e}"
             )
+
+    def shutdown(self, signum=None, frame=None, exit_code=0):
+        """Handle shutdown process by stopping all running processes and cleaning up."""
+        self.shutting_down = True
+        self.logger.info("Shutdown signal received. Cleaning up...")
+        processes_to_stop = [
+            "riven_frontend",
+            "riven_backend",
+            "Zilean",
+            "PostgreSQL",
+            "Zurg",
+            "rclone",
+            "pgAdmin",
+            "pgAgent",
+        ]
+        for process_name in processes_to_stop:
+            self.stop_process(process_name)
+        self.unmount_all()
+        self.logger.info("Shutdown complete.")
+        sys.exit(exit_code)
+
+    def unmount_all(self):
+        """Unmount all mount points in a specified directory."""
+        RCLONEDIR = "/path/to/rclone/mounts"
+        for mount_point in os.listdir(RCLONEDIR):
+            full_path = os.path.join(RCLONEDIR, mount_point)
+            if os.path.ismount(full_path):
+                self.logger.info(f"Unmounting {full_path}...")
+                umount = subprocess.run(
+                    ["umount", full_path], capture_output=True, text=True
+                )
+                if umount.returncode == 0:
+                    self.logger.info(f"Successfully unmounted {full_path}")
+                else:
+                    self.logger.error(
+                        f"Failed to unmount {full_path}: {umount.stderr.strip()}"
+                    )